requirements.txt
pypdf>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
llama-parse>=0.1.0
llama-index-core>=0.10.0
llama-index-embeddings-openai>=0.1.0
numpy>=1.24.0